Added hard constrained cluster with hoppers.
The customk file can be used to run through hopping files generated with
hop_gen. These results are saved via a dataframe into a csv file for
ease of seaborn plotting. Currently if a point is not assigned it is
force assigned which should be changed when observing points not
assigned relative to constraint number.